bsd-user: move arch/OS dependent code out of main.c
This change moves the cpu initialization and main loop code from
main.c to the OS and arch dependent directories. This eliminates
many of the #ifdef's in main.c. The cpu initialization and loop
code is now located in the arch directory along with target arch
support code.

Signed-off-by: Stacey Son <sson@FreeBSD.org>
Signed-off-by: Sean Bruno <sbruno@FreeBSD.org>
Signed-off-by: Juergen Lock <nox@jelal.kn-bremen.de>
Signed-off-by: Michal Meloun <mmel@FreeBSD.org>
[imp integrated API change and functioanl fixes from mmel, sbruno and nox]
Signed-off-by: Warner Losh <imp@FreeBSD.org>

Fix typo in  fork_end() in bsd-user/main.c.
If emulated program starts new thread first and
then do fork(), then qemu fails with SISEGV.
This issue can be observed in python scons or Waf.