bsd-user: Add target coredump file generation.
This code change adds a feature that writes an ELF coredump file out
when the target code gets an uncaught signal that has the default
action of dumping a core file.  The coredump with be written to
qemu_<basename_of_target_binary>.core.  The coredump file then can be
used with a target or a cross debugging tool such as gdb, lldb,
procstat, etc. to examine the cause of the crash.

With slight rework by Ed Schouten <ed@nuxi.nl> to remove basename_r
and some strdups.